api/requirements.txt
fastapi>=0.109,<1.0
uvicorn[standard]>=0.27,<1.0
gspread>=6.0,<7.0
google-auth>=2.28,<3.0
python-dotenv>=1.0,<2.0
pandas>=2.2,<3.0
